lpc55_isp: refactor with trait
