+++ lpc55_isp/src/cmd.rs
@@ -10,27 +10,27 @@ enum DataPhase<'a> {
     Recv { code: ResponseCode, cnt: u32 },
 }
 
-fn do_command(
-    port: &mut dyn serialport::SerialPort,
+fn do_command<P: Isp>(
+    port: &mut P,
     tag: CommandTag,
     command_resp: ResponseCode,
     args: impl Into<Vec<u32>>,
     d: DataPhase,
 ) -> Result<Option<Vec<u8>>, IspError> {
-    send_command(port, tag, args)?;
+    port.send_command(tag, args)?;
 
-    read_response(port, command_resp)?;
+    port.read_response(command_resp)?;
 
     let ret = match d {
         DataPhase::NoData => None,
         DataPhase::Send { code, data } => {
-            send_data(port, data)?;
-            read_response(port, code)?;
+            port.send_data(data)?;
+            port.read_response(code)?;
             None
         }
         DataPhase::Recv { code, cnt } => {
-            let r = recv_data(port, cnt)?;
-            read_response(port, code)?;
+            let r = port.recv_data(cnt)?;
+            port.read_response(code)?;
             Some(r)
         }
     };
@@ -38,7 +38,7 @@ fn do_command(
     Ok(ret)
 }
 
-pub fn do_save_keystore(port: &mut dyn serialport::SerialPort) -> Result<(), IspError> {
+pub fn do_save_keystore<P: Isp>(port: &mut P) -> Result<(), IspError> {
     do_command(
         port,
         CommandTag::KeyProvision,
@@ -55,7 +55,7 @@ pub fn do_save_keystore(port: &mut dyn serialport::SerialPort) -> Result<(), Isp
     Ok(())
 }
 
-pub fn do_enroll(port: &mut dyn serialport::SerialPort) -> Result<(), IspError> {
+pub fn do_enroll<P: Isp>(port: &mut P) -> Result<(), IspError> {
     do_command(
         port,
         CommandTag::KeyProvision,
@@ -67,7 +67,7 @@ pub fn do_enroll(port: &mut dyn serialport::SerialPort) -> Result<(), IspError> 
     Ok(())
 }
 
-pub fn do_generate_uds(port: &mut dyn serialport::SerialPort) -> Result<(), IspError> {
+pub fn do_generate_uds<P: Isp>(port: &mut P) -> Result<(), IspError> {
     do_command(
         port,
         CommandTag::KeyProvision,
@@ -86,10 +86,7 @@ pub fn do_generate_uds(port: &mut dyn serialport::SerialPort) -> Result<(), IspE
     Ok(())
 }
 
-pub fn do_isp_write_keystore(
-    port: &mut dyn serialport::SerialPort,
-    data: &[u8],
-) -> Result<(), IspError> {
+pub fn do_isp_write_keystore<P: Isp>(port: &mut P, data: &[u8]) -> Result<(), IspError> {
     do_command(
         port,
         CommandTag::KeyProvision,
@@ -104,7 +101,7 @@ pub fn do_isp_write_keystore(
     Ok(())
 }
 
-pub fn do_recv_sb_file(port: &mut dyn serialport::SerialPort, data: &[u8]) -> Result<(), IspError> {
+pub fn do_recv_sb_file<P: Isp>(port: &mut P, data: &[u8]) -> Result<(), IspError> {
     do_command(
         port,
         CommandTag::ReceiveSbFile,
@@ -119,8 +116,8 @@ pub fn do_recv_sb_file(port: &mut dyn serialport::SerialPort, data: &[u8]) -> Re
     Ok(())
 }
 
-pub fn do_isp_set_userkey(
-    port: &mut dyn serialport::SerialPort,
+pub fn do_isp_set_userkey<P: Isp>(
+    port: &mut P,
     key_type: KeyType,
     data: &[u8],
 ) -> Result<(), IspError> {
@@ -145,8 +142,8 @@ pub fn do_isp_set_userkey(
     Ok(())
 }
 
-pub fn do_isp_read_memory(
-    port: &mut dyn serialport::SerialPort,
+pub fn do_isp_read_memory<P: Isp>(
+    port: &mut P,
     address: u32,
     cnt: u32,
 ) -> Result<Vec<u8>, IspError> {
@@ -169,8 +166,8 @@ pub fn do_isp_read_memory(
     Ok(f.unwrap())
 }
 
-pub fn do_isp_write_memory(
-    port: &mut dyn serialport::SerialPort,
+pub fn do_isp_write_memory<P: Isp>(
+    port: &mut P,
     address: u32,
     data: &[u8],
 ) -> Result<(), IspError> {
@@ -195,7 +192,7 @@ pub fn do_isp_write_memory(
     Ok(())
 }
 
-pub fn do_isp_flash_erase_all(port: &mut dyn serialport::SerialPort) -> Result<(), IspError> {
+pub fn do_isp_flash_erase_all<P: Isp>(port: &mut P) -> Result<(), IspError> {
     do_command(
         port,
         CommandTag::FlashEraseAll,
@@ -210,8 +207,8 @@ pub fn do_isp_flash_erase_all(port: &mut dyn serialport::SerialPort) -> Result<(
     Ok(())
 }
 
-pub fn do_isp_flash_erase_region(
-    port: &mut dyn serialport::SerialPort,
+pub fn do_isp_flash_erase_region<P: Isp>(
+    port: &mut P,
     start_address: u32,
     byte_count: u32,
 ) -> Result<(), IspError> {
@@ -231,20 +228,19 @@ pub fn do_isp_flash_erase_region(
     Ok(())
 }
 
-pub fn do_isp_get_property(
-    port: &mut dyn serialport::SerialPort,
+pub fn do_isp_get_property<P: Isp>(
+    port: &mut P,
     prop: BootloaderProperty,
 ) -> Result<Vec<u32>, IspError> {
-    send_command(port, CommandTag::GetProperty, [prop as u32])?;
+    port.send_command(CommandTag::GetProperty, [prop as u32])?;
 
-    let f = read_response(port, ResponseCode::GetProperty)?;
+    let f = port.read_response(ResponseCode::GetProperty)?;
 
     Ok(f)
 }
 
-pub fn do_isp_last_error(port: &mut dyn serialport::SerialPort) -> Result<Vec<u32>, IspError> {
-    send_command(
-        port,
+pub fn do_isp_last_error<P: Isp>(port: &mut P) -> Result<Vec<u32>, IspError> {
+    port.send_command(
         CommandTag::GetProperty,
         [
             // Arg 0 = LastCRC
@@ -254,7 +250,7 @@ pub fn do_isp_last_error(port: &mut dyn serialport::SerialPort) -> Result<Vec<u3
         ],
     )?;
 
-    let f = read_response(port, ResponseCode::GetProperty)?;
+    let f = port.read_response(ResponseCode::GetProperty)?;
 
     Ok(f)
 }
