+++ lpc55_isp/src/lib.rs
@@ -4,3 +4,4 @@
 
 pub mod cmd;
 pub mod isp;
+pub mod serial;
